2016/CVE-2016-4631/requirements.txt
# Dependencias para FuxIOS.py v2.0
# CVE-2016-4631 Exploit PoC

scapy>=2.4.5
python-nmap>=0.7.1

# Dependencias adicionales que pueden ser necesarias
# En algunos sistemas puede requerirse:
# netifaces>=0.11.0
# pycryptodome>=3.15.0

# Nota: También se requiere tener instalado nmap en el sistema
# Windows: https://nmap.org/download.html
# Linux: sudo apt-get install nmap
# macOS: brew install nmap
